Extract EncryptedResource decrypters to separate files
